Deliver my pernicek you dirty courier

+++ product-selection-service/src/main/kotlin/org/pv293/kotlinseminar/productSelectionService/application/commands/impl/ChooseLocationCommand.kt
@@ -1,11 +1,12 @@
 package org.pv293.kotlinseminar.productSelectionService.application.commands.impl
 
 import org.axonframework.modelling.command.TargetAggregateIdentifier
+import java.math.BigDecimal
 import java.util.UUID
 
 data class ChooseLocationCommand(
     @TargetAggregateIdentifier
     val locationId: UUID,
-    val latitude: Double,
-    val longitude: Double,
+    val latitude: BigDecimal,
+    val longitude: BigDecimal,
 )
